requirements.txt
beautifulsoup4
cachecontrol[filecache]
click
python-dateutil
requests
stevedore
toml
